:wrench: dynamic reflection of nfts_pallet
